Changed resolution

+++ TRLogicApp/Services/ImageWorker.cs
@@ -45,7 +45,7 @@ namespace TRLogicApp.Services
             var destRect = new Rectangle(0, 0, width, height);
             var destImage = new Bitmap(width, height);
 
-            destImage.SetResolution(image.HorizontalResolution, image.VerticalResolution);
+            destImage.SetResolution((int)image.HorizontalResolution, (int)image.VerticalResolution);
 
             using (var graphics = Graphics.FromImage(destImage))
             {
